% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{CaDrA_Server}
\alias{CaDrA_Server}
\title{Shiny Server modules}
\usage{
CaDrA_Server(id, datalist_file = NULL)
}
\arguments{
\item{id}{A unique namespace identifier}

\item{datalist_file}{A path to data file (in cvs or rds format) listing 
the absolute paths of necessary files to start the app. 

The datalist_file must contains the following variables:
'feature_set' (required), 'feature_set_paths' (repuired), 
'input_score_names' (required), input_score_paths' (required), 
'gene_expression_names' (optional), 'gene_expression_paths' (optional)

Default is NULL. If NULL, the app will start with DEFAULT dataset included 
in CaDrA package.}
}
\value{
Shiny server elements
}
\description{
Shiny Server modules
}
\examples{

# Load R library
library(shiny)

# Create ui and server for Shiny app
id <- "myapp"

ui <- shiny::fluidPage(
   CaDrA-shiny::CaDrA_UI(id = id)
)

server <- function(input, output, session) {
   CaDrA-shiny::CaDrA_Server(id = id)
}

app <-  shiny::shinyApp(ui=ui, server=server)

# Launch and deploy Shiny app (NOT RUN)
# shiny::runApp(app, host='0.0.0.0', port=3838)
 
}
