% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{CaDrA_UI}
\alias{CaDrA_UI}
\title{Shiny UI modules}
\usage{
CaDrA_UI(id, datalist = NULL)
}
\arguments{
\item{id}{A unique namespace identifier}

\item{datalist}{A data frame or path to data file (in cvs or rds format) listing 
the absolute paths of necessary files to start the app. 

The datalist must contains the following variables:
'feature_set' (required), 'feature_set_path' (repuired), 
'input_score_name' (required), input_score_path' (required), 
'gene_expression_name' (optional), 'gene_expression_path' (optional)

Default is NULL. If NULL, the app will start with DEFAULT dataset included 
in CaDrA package.}
}
\value{
shiny ui elements
}
\description{
Shiny UI modules
}
\examples{

# Load R library
library(shiny)

# Create ui and server for Shiny app
id <- "myapp"

ui <- shiny::fluidPage(
   CaDrA.shiny::CaDrA_UI(id = id)
)

server <- function(input, output, session) {
   CaDrA.shiny::CaDrA_Server(id = id)
}

app <- shiny::shinyApp(ui=ui, server=server)

# Launch Shiny app (NOT RUN)
# shiny::runApp(app, host='0.0.0.0', port=3838)

}
