% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_api.R
\name{get_feature_set}
\alias{get_feature_set}
\title{Get a list of pre-processed feature sets available on CaDrA API Portal}
\usage{
get_feature_set(order_by = "asc")
}
\arguments{
\item{order_by}{sort results by ascending (asc) or descending order (desc)}
}
\value{
a data frame with two columns: description, feature_set_name
}
\description{
Get a list of pre-processed feature sets available on CaDrA API Portal
}
\examples{

# Retrieve a list of feature sets available from CaDrA Portal
fs_list <- CaDrA.shiny::get_feature_set(
  order_by = "asc"
)

# View top 6 rows
head(fs_list)

}
