% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_api.R
\name{get_feature_sets}
\alias{get_feature_sets}
\title{Get a list of pre-processed feature sets available on CaDrA's API Portal}
\usage{
get_feature_sets(API_Server, orderby = "asc")
}
\arguments{
\item{API_Server}{an URL link to where the API is hosted}

\item{orderby}{sort results by ascending or descending order}
}
\value{
a data frame with two columns: description, feature_set_names
}
\description{
Get a list of pre-processed feature sets available on CaDrA's API Portal
}
\examples{

# Load R library
# examples will go here

}
