% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_api.R
\name{get_gene_expression}
\alias{get_gene_expression}
\title{Get a list of gene expression sets associated with a given feature set}
\usage{
get_gene_expression(API_Server, feature_set, order_by = "asc")
}
\arguments{
\item{API_Server}{an URL link to where the API is hosted}

\item{feature_set}{a specific feature set to look up and retrieve 
appropriate gene expression sets that are associated with it}

\item{order_by}{sort results by ascending or descending order}
}
\value{
a data frame with two columns: feature_set_name, gene_expression_name
}
\description{
Get a list of gene expression sets associated with a given feature set
}
\examples{
 
# API server to retrieve a gene expression set that associated with a 
specific feature set from CaDrA Portal
ge_API_Server <- "https://montilab.bu.edu/CaDrA_API/get_gene_expression"

ge_list <- CaDrA.shiny::get_gene_expression(
  API_Server = ge_API_Server, 
  feature_set = "TCGA_ACC_2016_01_28_GISTIC_MUT_SIG", 
  order_by = "asc"
)

}
