% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_api.R
\name{download_feature_sets}
\alias{download_feature_sets}
\title{Get a specific feature set available on CaDrA API Portal
and determine whether to include input scores and gene expression that 
associated with that feature set}
\usage{
download_feature_sets(
  feature_set,
  include_input_score = TRUE,
  include_gene_expression = TRUE,
  out_dir = Sys.getenv("HOME")
)
}
\arguments{
\item{feature_set}{a list of feature sets to look up and download the 
appropriate data that are associated with it.}

\item{include_input_score}{a Boolean value to determine whether to include 
input scores that associated with the given feature sets. Default is TRUE.}

\item{include_gene_expression}{a Boolean value to determine whether to 
include gene expression sets that associated with the given feature sets. 
Default is TRUE.}

\item{out_dir}{a directory to save the output.}
}
\value{
a zip file
}
\description{
Get a specific feature set available on CaDrA API Portal
and determine whether to include input scores and gene expression that 
associated with that feature set
}
\examples{

library(shiny)

# Retrieve a list of feature sets available from CaDrA Portal
fs_list <- CaDrA.shiny::get_feature_set(order_by = "asc")

# Download feature sets and return a datalist with appropriate paths to its dataset
mydatafile <- download_feature_sets(
  feature_set = fs_list$feature_set_name[1],
  include_input_score = TRUE,
  include_gene_expression = TRUE,
  out_dir = "~/Github"
)

# Launch CaDrA Shiny app with your downloaded datalist
app <- CaDrA.shiny::CaDrA_App(id="myapp", datalist=mydatafile)

# Launch Shiny app at localhost with port 3838 (NOT RUN)
# shiny::runApp(app, host='0.0.0.0', port=3838)

}
