% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_api.R
\name{download_feature_sets}
\alias{download_feature_sets}
\title{Get a specific feature set available on CaDrA's API Portal
and determine whether to include input scores and gene expression that 
associated with that feature set}
\usage{
download_feature_sets(
  API_Server,
  feature_sets,
  include_input_score = TRUE,
  include_gene_expression = TRUE,
  out_dir
)
}
\arguments{
\item{API_Server}{an URL link to where the API is hosted}

\item{feature_sets}{a list of feature sets to look up and download the 
appropriate data that are associated with it.}

\item{include_input_score}{a Boolean value to determine whether to include 
input scores that associated with the given feature sets. Default is TRUE.}

\item{include_gene_expression}{a Boolean value to determine whether to 
include gene expression sets that associated with the given feature sets. 
Default is TRUE.}

\item{out_dir}{a directory to save the output.}
}
\value{
a zip file
}
\description{
Get a specific feature set available on CaDrA's API Portal
and determine whether to include input scores and gene expression that 
associated with that feature set
}
\examples{

# Load R library
# examples will go here

}
