% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_api.R
\name{get_input_score}
\alias{get_input_score}
\title{Get a list of input scores associated with a given feature set}
\usage{
get_input_score(API_Server, feature_set, orderby = "asc")
}
\arguments{
\item{API_Server}{an URL link to where the API is hosted}

\item{feature_set}{a specific feature set to look up and retrieve 
appropriate input scores that are associated with it}

\item{orderby}{sort results by ascending or descending order}
}
\value{
a data frame with two columns: feature_set_names, input_score_names
}
\description{
Get a list of input scores associated with a given feature set
}
\examples{

# Load R library
# examples will go here

}
