% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadra_api.R
\name{pull_datasets}
\alias{pull_datasets}
\title{Pull a specific feature set available on CaDrA Portal
and determine whether to include gene expression set that is 
associated with that feature set}
\usage{
pull_datasets(feature_set, include_gene_expression = FALSE)
}
\arguments{
\item{feature_set}{a list of feature sets to look up and download the 
appropriate data that are associated with it.}

\item{include_gene_expression}{a Boolean value to determine whether to 
include gene expression sets that associated with the given feature sets. 
Default is TRUE.}
}
\value{
a list of feature set and gene expression set if 
include_gene_expression = TRUE
}
\description{
Pull a specific feature set available on CaDrA Portal
and determine whether to include gene expression set that is 
associated with that feature set
}
\examples{

# Load R library
library(shiny)

# Retrieve a list of feature sets available from CaDrA Portal
fs_list <- CaDrA.shiny::get_feature_set(order_by = "asc")

# Download feature sets and return a datalist with appropriate paths to its dataset
datasets <- pull_datasets(
  feature_set = fs_list$feature_set_name[1],
  include_gene_expression = TRUE
)

}
